+++ pkg/auth/api_key.go
@@ -5,8 +5,9 @@ import (
 )
 
 const (
-	GenerateAPIKeyEndpoint       = "system/generate-api-key"
-	GetDatasourcesAPIKeyEndpoint = "system/get-datasources-api-key"
+	GenerateAPIKeyEndpoint            = "system/generate-api-key"
+	GetDatasourcesAPIKeyEndpoint      = "system/get-datasources-api-key"
+	GenerateClientTokenAPIKeyEndpoint = "system/generate-client-token"
 )
 
 type ApiKey struct {